+++ mall-system/src/main/java/org/example/modules/admin/order/controller/OrderReturnApplyController.java
@@ -1,7 +1,7 @@
 package org.example.modules.admin.order.controller;
 
 
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.Resource;
@@ -18,6 +18,7 @@ import org.example.security.annotaion.rest.AnonymousDeleteMapping;
 import org.example.security.annotaion.rest.AnonymousGetMapping;
 import org.example.security.annotaion.rest.AnonymousPostMapping;
 import org.example.security.annotaion.rest.AnonymousPutMapping;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -25,6 +26,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.io.Serializable;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -67,10 +69,13 @@ public class OrderReturnApplyController {
      * @return 单条数据
      */
     @AnonymousGetMapping("{id}")
-    public ResponseEntity<Object> selectOne(@PathVariable Serializable id) {
+    public ResponseEntity<Object> selectOne(@PathVariable @NotNull Serializable id) {
         // TODO 优化代码
         OrderReturnApplyEntity orderReturnApplyEntity = this.orderReturnApplyService.getById(id);
         OrderReturnApplyVo orderReturnApplyVo = BeanCopy.convert(orderReturnApplyEntity, OrderReturnApplyVo.class);
+        if (Objects.isNull(orderReturnApplyVo)) {
+            return ResponseEntity.ok("传入正确的id");
+        }
         CompanyAddressEntity companyAddressEntity = companyAddressService.getById(orderReturnApplyVo.getCompanyAddressId());
         CompanyAddressVo companyAddressVo = BeanCopy.convert(companyAddressEntity, CompanyAddressVo.class);
         // 退货地址
@@ -116,6 +121,9 @@ public class OrderReturnApplyController {
      */
     @AnonymousDeleteMapping
     public ResponseEntity<Object> remove(@RequestBody Set<Long> idList) {
+        if (CollectionUtils.isEmpty(idList)) {
+            throw new BaseRequestException("请正确的填写id");
+        }
         return ResponseEntity.ok(this.orderReturnApplyService.removeByIds(idList.stream().filter(id -> String.valueOf(id).length() < 20 && String.valueOf(id).length() >= 1).limit(10).collect(Collectors.toSet())) ? "删除成功" : "删除失败");
     }
 }
