+++ mall-system/src/main/java/org/example/modules/admin/product/controller/ProductAttributeController.java
@@ -1,6 +1,7 @@
 package org.example.modules.admin.product.controller;
 
 
+import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.Resource;
@@ -110,6 +111,9 @@ public class ProductAttributeController {
      */
     @AnonymousDeleteMapping
     public ResponseEntity<Object> remove(@RequestBody Set<Long> idList) {
+        if (CollectionUtils.isEmpty(idList)) {
+            throw new BaseRequestException("请正确的填写id");
+        }
         return ResponseEntity.ok(this.productAttributeService.removeByIds(idList.stream().filter(id -> String.valueOf(id).length() < 20 && String.valueOf(id).length() >= 1).limit(10).collect(Collectors.toSet())) ? "删除成功" : "删除失败");
     }
 }
