+++ mall-system/src/main/java/org/example/modules/admin/product/controller/SkuStockController.java
@@ -2,9 +2,11 @@ package org.example.modules.admin.product.controller;
 
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.annotation.Resource;
+import org.example.common.core.exception.BaseRequestException;
 import org.example.modules.admin.product.entity.SkuStockEntity;
 import org.example.modules.admin.product.entity.dto.SkuStockDto;
 import org.example.modules.admin.product.service.SkuStockService;
@@ -99,7 +101,10 @@ public class SkuStockController {
      * @return 删除结果
      */
     @AnonymousDeleteMapping
-    public ResponseEntity<Object> remove(@RequestBody Set<Long> idList) {
+        public ResponseEntity<Object> remove(@RequestBody Set<Long> idList) {
+        if (CollectionUtils.isEmpty(idList)) {
+            throw new BaseRequestException("请正确的填写id");
+        }
         return ResponseEntity.ok(this.skuStockService.removeByIds(idList.stream().filter(id -> String.valueOf(id).length() < 20 && String.valueOf(id).length() >= 1).limit(10).collect(Collectors.toSet())) ? "删除成功" : "删除失败");
     }
 }
